fix: compatibility with zod v4 and zod-openapi v5
- Update tests to reflect newest zod v4
- Add legacy overrides script to support zod v3 + zod-openapi v4 as well as zod v4 + zod-openapi v5
- Fix `components` schema merging
