+++ pnpm-workspace.disabled
@@ -1,0 +1,3 @@
+overrides:
+  zod: npm:zod@3.25.76
+  zod-openapi: npm:zod-openapi@4.2.4
